+++ config/splat.ric.yaml
@@ -32,6 +32,9 @@ segments:
       - [0x19EE0, rodata]
       - [0x1A454, rodata]
       - [0x1AB04, rodata]
-      - [0x1AC60, c]
+      - [0x1AB68, .rodata, 32324]
+      - [0x1AB88, rodata]
+      - [0x1AC60, c, 1AC60]
+      - [0x32324, c, 32324]
       - [0x37B60, data]
   - [0x39A58]
