+++ src/ric/32324.c
@@ -1,0 +1,143 @@
+#include "common.h"
+#include "game.h"
+#include "objects.h"
+#include "sfx.h"
+
+void func_8016E324(Entity* entity) {
+    switch (entity->step) {
+    case 0:
+        entity->unk34 = 0x04020000;
+        entity->unk7C.s = 0x14;
+        entity->step++;
+
+    case 1:
+        entity->unk7C.s--;
+        if ((entity->unk7C.s) == 0) {
+        case 3:
+        case 5:
+            func_801606BC(entity, 0x39, 0);
+            entity->step++;
+        case 2:
+        case 4:
+        case 6:
+            entity->unk7C.s++;
+            if (entity->unk7C.s >= 11) {
+                entity->unk7C.s = 0;
+                entity->posX.val = 0x800000;
+                entity->posY.val = 0;
+                func_801606BC(entity, 0x10004, 0);
+                entity->step++;
+            }
+        }
+        break;
+
+    case 7:
+        entity->unk7C.s++;
+        if (entity->unk7C.s >= 16) {
+            func_80156C60(entity);
+            D_80072F6E = 1;
+            func_801606BC(entity, 0x3A, 0);
+        }
+        break;
+    }
+}
+
+INCLUDE_ASM("asm/ric/nonmatchings/32324", func_8016E46C);
+
+INCLUDE_ASM("asm/ric/nonmatchings/32324", func_8016E9E4);
+
+INCLUDE_ASM("asm/ric/nonmatchings/32324", func_8016F198);
+
+INCLUDE_ASM("asm/ric/nonmatchings/32324", func_8016F6F0);
+
+void func_80170548(Entity* entity) {
+    switch (entity->step) {
+    case 0:
+        entity->unk34 = 0x04000000;
+        entity->unkB0 = 0x1E;
+        func_8015FAB8(entity);
+        entity->hitboxWidth = 8;
+        entity->hitboxHeight = 8;
+        entity->step++;
+        break;
+
+    case 1:
+        if (++entity->unk7C.s >= 5) {
+            func_80156C60(entity);
+        }
+        break;
+    }
+}
+
+INCLUDE_ASM("asm/ric/nonmatchings/32324", func_801705EC);
+
+INCLUDE_ASM("asm/ric/nonmatchings/32324", func_801706C0);
+
+INCLUDE_ASM("asm/ric/nonmatchings/32324", func_80170874);
+
+INCLUDE_ASM("asm/ric/nonmatchings/32324", func_8017091C);
+
+INCLUDE_ASM("asm/ric/nonmatchings/32324", func_80170F64);
+
+INCLUDE_ASM("asm/ric/nonmatchings/32324", func_8017161C);
+
+INCLUDE_ASM("asm/ric/nonmatchings/32324", func_801719A4);
+
+void func_80172AE8(Entity* entity) {
+    POLY_GT4* poly;
+    s32 ret;
+
+    switch (entity->step) {
+    case 0:
+        ret = g_api.AllocPolygons(4, 1);
+        entity->firstPolygonIndex = ret;
+        if (entity->firstPolygonIndex != -1) {
+            entity->unk34 = 0x04820000;
+            poly = &D_80086FEC[entity->firstPolygonIndex];
+            poly->tpage = 0x1C;
+            poly->clut = 0x19D;
+            poly->u2 = 0x20;
+            poly->u0 = 0x20;
+            poly->u3 = 0x30;
+            poly->u1 = 0x30;
+            poly->v1 = 0;
+            poly->v0 = 0;
+            poly->v3 = 0x10;
+            poly->v2 = 0x10;
+            poly->x0 = poly->x2 = entity->posX.i.hi - 8;
+            poly->x1 = poly->x3 = entity->posX.i.hi + 8;
+            poly->y0 = poly->y1 = entity->posY.i.hi - 8;
+            poly->y2 = poly->y3 = entity->posY.i.hi + 8;
+            poly->pad2 = entity->zPriority;
+            poly->pad3 = 0x115;
+            entity->unk7E.modeU16 = 0x60U;
+            entity->step++;
+        } else {
+            func_80156C60(entity);
+            return;
+        }
+        break;
+
+    case 1:
+        if (++entity->unk7C.s > 5) {
+            entity->step++;
+        }
+        entity->unk7E.modeU16 -= 8;
+        break;
+
+    case 2:
+        func_80156C60(entity);
+        return;
+
+    default:
+        break;
+    }
+    poly = &D_80086FEC[entity->firstPolygonIndex];
+    poly->r0 = poly->r1 = poly->r2 = poly->r3 = poly->g0 = poly->g1 = poly->g2 =
+        poly->g3 = poly->b0 = poly->b1 = poly->b2 = poly->b3 =
+            entity->unk7E.modeU8.unk0;
+}
+
+INCLUDE_ASM("asm/ric/nonmatchings/32324", func_80172D00);
+
+INCLUDE_ASM("asm/ric/nonmatchings/32324", func_8017328C);
